requirements.txt
alembic==1.6.3
appdirs==1.4.4
asttokens==2.0.5
attrs==21.2.0
black==21.5b1
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.14.1
executing==0.6.0
flasgger==0.9.5
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
icecream==2.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
mistune==0.8.4
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.2.0
PyYAML==5.4.1
regex==2021.4.4
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.24
toml==0.10.2
Werkzeug==2.0.1
